feat: introduce `EmailServiceTest` as wrapper for email service tests
