+++ src/test/kotlin/se/uulm/snowballr/backend/service/email/SendVerificationEmailTest.kt
@@ -1,8 +1,5 @@
 package se.uulm.snowballr.backend.service.email
 
-import com.icegreen.greenmail.configuration.GreenMailConfiguration
-import com.icegreen.greenmail.junit5.GreenMailExtension
-import com.icegreen.greenmail.util.ServerSetupTest
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
@@ -11,13 +8,9 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import org.junit.jupiter.api.extension.RegisterExtension
-import org.simplejavamail.MailException
-import org.simplejavamail.api.mailer.Mailer
 import org.simplejavamail.api.mailer.config.TransportStrategy
 import org.simplejavamail.mailer.MailerBuilder
 import se.uulm.snowballr.backend.env.Env
-import se.uulm.snowballr.backend.env.EnvReader
 import se.uulm.snowballr.backend.mail.EmailTemplateManager
 import se.uulm.snowballr.backend.model.SnowballRException
 import se.uulm.snowballr.backend.model.SnowballRException.EmailException
@@ -25,27 +18,7 @@ import se.uulm.snowballr.backend.model.email.EmailData
 import se.uulm.snowballr.backend.model.email.EmailTemplate
 import se.uulm.snowballr.backend.service.EmailService
 
-class SendVerificationEmailTest {
-    private val testSenderName = "SnowballR Test"
-    private val testSenderEmail = "noreply@snowballr.test"
-    private val testFrontendUrl = "https://frontend.test"
-    private val recipientEmail = "test.user@example.com"
-
-    private val envReaderMock = mockk<EnvReader>()
-    private val mailerMock = mockk<Mailer>()
-    private val emailTemplateManagerMock = mockk<EmailTemplateManager>()
-
-    // Helper class to throw a custom exception from within the mailer.
-    private class TestMailException(message: String, cause: Throwable? = null) : MailException(message, cause)
-
-    companion object {
-        @JvmField
-        @RegisterExtension
-        val greenMail: GreenMailExtension = GreenMailExtension(ServerSetupTest.SMTP)
-            .withConfiguration(GreenMailConfiguration.aConfig().withUser("user", "pass"))
-            .withPerMethodLifecycle(true)
-    }
-
+class SendVerificationEmailTest : EmailServiceTest() {
     @BeforeEach
     fun setUp() {
         val miscellaneousMock = mockk<Env.Miscellaneous>()
